requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
flask
gunicorn
torch
torchvision
pillow
numpy